[ch02] Added comments along with a little fix
The code may be fine. But there is NO inline comments, which makes it
very hard to decipher. Some explanatory comments are added.

In addition, one little fix is made: the input data path is specified as
the 1st argument, instead of being hard-coded. Now we could run the code
as:
spark-submit --master spark://localhost.localdomain:7077 \
  --executor-memory 512M --class com.cloudera.datascience.intro.RunIntro \
  target/ch02-intro-1.0.2.jar hdfs://localhost:9000/user/r/linkage